fix: real CS1503 in MigrationService.RollbackAsync (long vs int)
Installed .NET 8 SDK (Ubuntu's dotnet-sdk-8.0 apt package) in this
remote session and ran a real dotnet build for the first time on this
branch. AMCCA.Core (plain net8.0, no WPF) compiled cleanly except for
one genuine error: RollbackAsync's MigrationsRequiringForeignKeysOff
.Contains(rec.Version) passed a long (MigrationRecord.Version) to a
HashSet<int>, which does not implicitly convert. UpgradeAsync's own
call one screen up already used an int (m.Version from
BuiltInMigrations) and compiled fine, which is why this went
unnoticed by manual reasoning alone. Cast to int at the call site;
migration versions are always small ints stored as long by SQLite.

AMCCA.Core now builds with 0 errors/0 warnings in this environment.
AMCCA.App and AMCCA.Core.Tests still cannot build here: both target
net8.0-windows with UseWPF=true, and Ubuntu's dotnet-sdk-8.0 package
does not ship Microsoft.NET.Sdk.WindowsDesktop at all (confirmed:
absent from /usr/lib/dotnet/sdk/8.0.130/Sdks). Microsoft's own SDK
CDN (builds.dotnet.microsoft.com, dotnetcli.azureedge.net) is blocked
by this session's egress policy, so the official SDK with WPF desktop
support cannot be fetched here either. Full build/test of AMCCA.App
and AMCCA.Core.Tests still requires the user's own Windows-capable
machine.

+++ src/AMCCA.Core/Database/MigrationService.cs
@@ -2356,7 +2356,7 @@ public class MigrationService
                 throw new InvalidOperationException($"No rollback definition for version {rec.Version}.");
             }
 
-            if (MigrationsRequiringForeignKeysOff.Contains(rec.Version))
+            if (MigrationsRequiringForeignKeysOff.Contains((int)rec.Version))
             {
                 // Mirrors the UpgradeAsync branch above: no ambient ADO transaction, since the
                 // DownSql text manages its own PRAGMA foreign_keys toggle and BEGIN/COMMIT.
